requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-factory-boy==1.0.0
django-phonenumber-field==7.0.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
ez-setup==0.9
factory-boy==3.2.1
Faker==15.3.2
gevent==22.10.2
greenlet==2.0.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
phonenumberslite==8.12.57
Pillow==9.3.0
prompt-toolkit==3.0.32
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.5.2
